pyproject.toml
[tool.poetry]
name = "lst-hack-parser"
version = "0.1.0"
description = "Cian.ru parser for lst-hack"
authors = ["Mirea Ninja <contact@mirea.ninja>"]

[tool.poetry.dependencies]
python = "^3.9"
alembic = "^1.8.1"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
python-dotenv = "^0.20.0"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"
pydantic = {extras = ["email"], version = "^1.9.2"}
aiohttp = "^3.8.1"
loguru = "^0.6.0"
requests = "^2.28.1"
geopy = "^2.2.0"
pandas = "^1.5.1"
numpy = "^1.23.4"
selenium = "^4.5.0"
webdriver-manager = "^3.8.4"
openpyxl = "^3.0.10"
packaging = "^21.3"


[tool.poetry.dev-dependencies]
mypy = "^0.942"
black = {extras = ["d"], version = "^22.8.0"}
isort = "^5.10.1"
autoflake = "^1.4"
flake8 = "^4.0.1"
pytest = "^7.1.1"


[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
